backend/requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
pycountry==24.6.1
python-dotenv==1.0.1
Requests==2.32.3
